+++ src/pages/Group/index.ts
@@ -1,3 +1,2 @@
-export { default } from './GroupList';
 export { default as GroupCreate } from './GroupCreate';
 export { default as GroupCreateSuccess } from './GroupCreateSuccess';
