+++ src/pages/My/index.ts
@@ -1,0 +1,3 @@
+export { default as GroupList } from './GroupList';
+export { default as BungaeList } from './BungaeList';
+export { default as Profile } from './Profile';
