+++ src/routes/routes.tsx
@@ -1,11 +1,9 @@
 import { ReactNode } from 'react';
 import ProetectedRoute from '@/routes/ProetectedRoute';
 import PublicRoute from '@/routes/PublicRoute';
-import GroupList, { GroupCreate, GroupCreateSuccess } from '@/pages/GroupList';
-import BungaeList from '@/pages/BungaeList';
-import MyPage from '@/pages/MyPage';
-import Login from '@/pages/Auth/Login';
-import LoginSuccess from '@/pages/Auth/LoginSuccess';
+import { Login, LoginSuccess } from '@/pages/Auth';
+import { BungaeList, GroupList, Profile } from '@/pages/My';
+import { GroupCreate, GroupCreateSuccess } from '@/pages/Group';
 
 interface RouteType {
   path: string;
@@ -16,7 +14,7 @@ interface RouteType {
 const proetectedRoutes: RouteType[] = [
   { path: '/', element: <GroupList />, layout: true },
   { path: '/mybungae', element: <BungaeList />, layout: true },
-  { path: '/mypage', element: <MyPage />, layout: true },
+  { path: '/myprofile', element: <Profile />, layout: true },
   { path: '/group/post', element: <GroupCreate />, layout: false },
   { path: '/group/success', element: <GroupCreateSuccess />, layout: false },
 ].map((route) => ({
